split make and revert snapshot into two seprate script files
